requirements.txt
agents==1.4.0
fastapi==0.116.1
mcp==1.13.1
pydantic==2.11.7
pydantic_settings==2.10.1
python-dotenv==1.1.1
